+++ stores/todoStore.ts
@@ -1,75 +1,20 @@
-import {
-  TodoItemProps,
-  TodoListProps,
-  TodoTemplateListProps,
-} from "@/types/Item";
+import { TodoItemProps, TodoListProps } from "@/types/Item";
 import { create } from "zustand";
 
 interface State {
-  todoTemplates: TodoTemplateListProps;
   todos: TodoListProps;
 }
 
 interface Action {
-  setTodoTemplates: (templates: TodoTemplateListProps) => void;
-  loadTodoTemplates: () => void;
-  handleSubmitTemplateItems: (e: React.FormEvent<HTMLFormElement>) => void;
-  deleteTemplateItems: (id: string) => void;
-  updateTemplateItems: (id: string, content: string) => void;
-  deleteTemplates: (templateId: string) => void;
   setTodos: (todos: TodoListProps) => void;
   loadTodos: () => void;
   handleSubmit: (e: React.FormEvent<HTMLFormElement>) => void;
-  updateTodoStatus: (id: string, mode: TodoItemProps["status"]) => void;
+  updateTodoItems: (id: string, content: string) => void;
+  deleteTodoItems: (id: string) => void;
+  toggleStatus: (id: string, mode: TodoItemProps["status"]) => void;
 }
 
 export const useTodoStore = create<State & Action>((set) => ({
-  // Templates
-  todoTemplates: [],
-  setTodoTemplates: (templates) => {
-    set({ todoTemplates: templates });
-  },
-  loadTodoTemplates: () => {
-    const storedData = localStorage.getItem("templates");
-    if (storedData) {
-      try {
-        set({ todoTemplates: JSON.parse(storedData) });
-      } catch (error) {
-        console.error("Failed to parse localStorage data:", error);
-        set({ todos: [] });
-      }
-    }
-  },
-  handleSubmitTemplateItems: (e) => {
-    e.preventDefault();
-    const formData = new FormData(e.currentTarget);
-    const todo = formData.get("todo")?.toString();
-
-    if (!todo?.trim()) return;
-
-    const newTemplate = {
-      id: new Date().getTime().toString(),
-      content: todo,
-      status: "start" as const,
-      priority: "low" as const,
-      // isSelected: false,
-    };
-
-    set((state) => {
-      const returnData = [...state.todoTemplates, newTemplate];
-
-      localStorage.setItem("templates", JSON.stringify(returnData));
-      return { todoTemplates: returnData };
-    });
-    e.currentTarget.reset();
-  },
-  deleteTemplateItems: (id) => {
-    set((state) => ({
-      todoTemplates: state.todoTemplates.filter((item) => item.id !== id),
-    }));
-  },
-  updateTemplateItems: (id, content) => {},
-  deleteTemplates(templateId) {},
   // Today todos
   todos: [],
   setTodos: (todos) => {
@@ -117,7 +62,32 @@ export const useTodoStore = create<State & Action>((set) => ({
     });
     e.currentTarget.reset();
   },
-  updateTodoStatus: (id, mode) =>
+  updateTodoItems: (id, content) => {
+    set((state) => {
+      const updatedTodos = state.todos.map((todo) =>
+        todo.id === id ? { ...todo, content } : todo
+      );
+
+      // today의 데이터만 업데이트하기 위해 필터링
+      const todayTodos = updatedTodos.filter((todo) => todo.date);
+      localStorage.setItem("today", JSON.stringify(todayTodos));
+
+      return { todos: updatedTodos };
+    });
+  },
+
+  deleteTodoItems: (id) => {
+    set((state) => {
+      const filteredTodos = state.todos.filter((todo) => todo.id !== id);
+
+      // today의 데이터만 저장하기 위해 필터링
+      const todayTodos = filteredTodos.filter((todo) => todo.date);
+      localStorage.setItem("today", JSON.stringify(todayTodos));
+
+      return { todos: filteredTodos };
+    });
+  },
+  toggleStatus: (id, mode) =>
     set((state) => ({
       todos: state.todos.map((todo) =>
         todo.id === id ? { ...todo, status: mode } : todo